% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{api_url}
\alias{api_url}
\title{URL setup helper}
\usage{
api_url(jc, endpoint)
}
\arguments{
\item{jc}{JATOS connection info, e.g. from \code{define_connection}}

\item{endpoint}{API endpoint address}
}
\value{
URL address created by matching the connection url and endpoint string.
}
\description{
URL setup helper
}
\examples{
\dontrun{
cc <- define_connection(
  "https://www.myjatosinstance.org",
  "Bearer jap_xXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXx"
)
# API for token info
api_url(cc, "/jatos/api/v1/admin/token")
}
}
