% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{base_request}
\alias{base_request}
\title{Basic JATOS HTTP request helper}
\usage{
base_request(jc, endpoint)
}
\arguments{
\item{jc}{JATOS connection info, e.g. from \code{define_connection}}

\item{endpoint}{API endpoint address}
}
\value{
\code{httr2::request} object
}
\description{
Basic JATOS HTTP request helper
}
\examples{
\dontrun{
cc <- define_connection(
  "https://www.myjatosinstance.org",
  "Bearer jap_xXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXx"
)

req <- base_request(cc, "/jatos/api/v1/admin/token")
}
}
