% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_date}
\alias{convert_date}
\title{Date conversion helper}
\usage{
convert_date(date_integer)
}
\arguments{
\item{date_integer}{Numerical format of times in JATOS metadata}
}
\value{
Lubridate date
}
\description{
Date conversion helper
}
\examples{
convert_date(1642617959901)
}
