% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_srt}
\alias{demo_srt}
\title{Simple Reaction Time Task raw response data}
\format{
\subsection{\code{demo_srt}}{

A list with two items:
\describe{
\item{meta}{\code{httr2_response} as returned by \code{get_metadata}}
\item{results}{\code{httr2_response} as returned by \code{get_results}, with
\code{body} referring to the bundled \code{srtdemo.zip} file,
which you can copy into current working directory with \code{prepare_demo}}
}
}
}
\source{
\url{https://github.com/visionlabels/jatosR}
}
\usage{
demo_srt
}
\description{
Simple reaction time study
(\href{https://github.com/JATOS/JATOS_examples/raw/main/examples/jspsych_7_simple_reaction_time_task.jzip}{Simple Reaction Time Task}),
which is included in \href{https://www.jatos.org/Example-Studies}{JATOS examples}.
The example study was uploaded to MindProbe server and run four times.
See \code{process_results} on how to turn the raw data into processed values.
}
\details{
This is the raw response data with two \code{httr2_response} values, as used
by \code{prepare_demo()}.
}
\keyword{datasets}
