% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_srt_processed.R
\docType{data}
\name{demo_srt_processed}
\alias{demo_srt_processed}
\title{Simple Reaction Time Task data}
\format{
\subsection{\code{who}}{

A list with two items:
\describe{
\item{meta}{A list with metadata}
\item{data}{A tibble with 4 rows and 26 columns. Raw data are in text format in the \code{raw} column.}
}
}
}
\source{
\url{https://github.com/visionlabels/jatosR}
}
\usage{
demo_srt_processed
}
\description{
Simple reaction time study
(\href{https://github.com/JATOS/JATOS_examples/raw/main/examples/jspsych_7_simple_reaction_time_task.jzip}{Simple Reaction Time Task}),
which is included in \href{https://www.jatos.org/Example-Studies}{JATOS examples}.
The example study was uploaded to MindProbe server and run four times.
You can play with the raw data yourself calling \code{prepare_demo} function.
The processed data are useful if you are interested
in the final format provided by jatosR package after calling \code{process_results}.
}
\keyword{datasets}
