% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{raw_jspsych}
\alias{raw_jspsych}
\title{Transforms raw data produced by jsPsych to tibble}
\usage{
raw_jspsych(d)
}
\arguments{
\item{d}{Tibble with data obtained with \code{process_results}}
}
\value{
Tibble with new list column \code{data_jspsych} or a list with \code{data} tibble (see above)
}
\description{
Takes raw results stored in \code{data_raw} and adds new column \code{data_jspsych} containing data in lists created by \code{jspsychread} package.
The packages \code{jsonlite} and \code{jspsychread} must be installed for this function to work.
It is expected that the parameter \code{d} is the dataset obtained with \code{process_results},
but the dataset can be simpler, because only the column \code{data_raw} is used.
}
\details{
Function \code{process_results} returns a list (let's assume called \code{results}),
\code{raw_jspsych} is intended to be used in tibble pipeline, i.e. on the \code{result$data} tibble.
To simplify the process, the function detects, whether it is provided
with a list containing a tibble called \code{data}. In that case,
it processes the \code{data} tibble, puts it back into the original list and returns it.
}
\examples{
\dontrun{
demo <- prepare_demo()
r <- process_results(demo$meta, demo$results)

# working with tibbles
rdata <- r$data |> raw_jspsych()
rdata |> select(start_date_sr, data_json)
rdata$data_json[[1]]

# shortcut for the results list also works
r <- r |> raw_jspsych()

clean_demo()
}
}
