% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_demo}
\alias{clean_demo}
\title{Demo clean up}
\usage{
clean_demo()
}
\value{
Nothing
}
\description{
Cleans temporary file after \code{prepare_demo()}.
}
\examples{
demo <- prepare_demo()
r <- process_results(demo$meta, demo$results)
r
clean_demo()
}
