% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{component_result_info}
\alias{component_result_info}
\title{Parser of componentResults data}
\usage{
component_result_info(cr)
}
\arguments{
\item{cr}{list corresponding to the componentResults part of the data}
}
\value{
List of size-one vectors, which could be converted to tibble if needed.
}
\description{
Parser of componentResults data
}
\examples{
demo <- prepare_demo(nofile = TRUE)
json <- demo$meta |> httr2::resp_body_json()
cr <- component_result_info(
  json$data[[1]]$studyResults[[1]]$componentResults[[1]])
cr
}
