% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{define_connection}
\alias{define_connection}
\title{Define JATOS connection}
\usage{
define_connection(url, token)
}
\arguments{
\item{url}{Base URL to the JATOS instance (with no API endpoints)}

\item{token}{API token obtained from JATOS instance

Check https://www.jatos.org/JATOS-API.html#personal-access-tokens
about the process of getting a token.

You don't need to worry about trailing slash in URL
(it is automatically removed). If you omit "Bearer " part
of the token, but the token seems to be valid, because it starts
with "jap", the "Bearer " part is added.}
}
\value{
A simple list with url and token values to be used in other functions
}
\description{
Define JATOS connection
}
\examples{
\dontrun{
cc <- define_connection(
  "https://www.myjatosinstance.org",
  "Bearer jap_xXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXx"
)
test_connection(cc)
}
}
