% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata for JATOS results}
\usage{
get_metadata(jc, batch_id = NULL, component_id = NULL)
}
\arguments{
\item{jc}{JATOS connection info, e.g. from \code{define_connection}}

\item{batch_id}{Integer identifier for a particular experiment batch}

\item{component_id}{Integer identifier for a particular experiment component

Only one of batch_id or component_id should be provided.}
}
\value{
HTTP response object returned by \code{httr2::req_perform()}. The actual metadata are stored in the response \code{body} and can be analyzed with \code{process_results}. The provided \code{batch_id} and \code{component_id} are added to the response data.
}
\description{
Get metadata for JATOS results
}
\examples{
\dontrun{
cc <- define_connection(
  "https://www.myjatosinstance.org",
  "Bearer jap_xXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXx"
)
meta <- get_metadata(cc, batch_id = 100)
meta$batch_id
}
}
