% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_results}
\alias{get_results}
\title{Get raw data for JATOS results}
\usage{
get_results(jc, batch_id = NULL, component_id = NULL, filename = NULL)
}
\arguments{
\item{jc}{JATOS connection info, e.g. from \code{define_connection}}

\item{batch_id}{Integer identifier for a particular experiment batch}

\item{component_id}{Integer identifier for a particular experiment component}

\item{filename}{Optional name of the file where the data should be stored. If not provided, temporary file is created in system folder for temporary files}
}
\value{
HTTP response object returned by \code{httr2::req_perform()}. The data are stored in the provided file and can be analyzed with \code{process_results}. The provided \code{batch_id} and \code{component_id} are added to the response data.
}
\description{
Get raw data for JATOS results
}
\examples{
\dontrun{
cc <- define_connection(
  "https://www.myjatosinstance.org",
  "Bearer jap_xXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXx"
)
results <- get_results(cc, batch_id = 100)
results$batch_id
# Name of the temporary file with the zipped data
results$body
}
}
