% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jatosR-package.R
\docType{package}
\name{jatosR-package}
\alias{jatosR}
\alias{jatosR-package}
\title{jatosR: Downloading Data From JATOS Servers}
\description{
The package helps with downloading data from a JATOS server. The main goal is to provide functions for authentication and basic processing for getting data to R formats. The package provides a thin wrapper around JATOS API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/visionlabels/jatosR}
  \item Report bugs at \url{https://github.com/visionlabels/jatosR/issues}
}

}
\author{
\strong{Maintainer}: Jiri Lukavsky \email{lukavsky@praha.psu.cas.cz} (\href{https://orcid.org/0000-0002-1082-229X}{ORCID})

}
\keyword{internal}
