% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_demo}
\alias{prepare_demo}
\title{Simple Response Time demo}
\usage{
prepare_demo(nofile = F)
}
\arguments{
\item{nofile}{Default \code{FALSE}. Use \code{TRUE} if you want to load only the metadata, without creating the local file.}
}
\value{
list with \code{meta} and \code{results}, which could be used in \code{process_results}. It also copies file \code{srt_demo.zip} in the working directory. After \code{process_results}, you can delete it, or use \code{clen_demo()}.
}
\description{
Simple Response Time demo
}
\examples{
demo <- prepare_demo()
r <- process_results(demo$meta, demo$results)
r
clean_demo()
}
