% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{process_results}
\alias{process_results}
\title{Turns metadata and data into a single table}
\usage{
process_results(meta, results, data_node = 1, component_node = 1)
}
\arguments{
\item{meta}{Metadata from \code{get_metadata} function}

\item{results}{Results from \code{get_results} function. It should link to an external file with zipped data.}

\item{data_node}{Default 1. You can change it if more data nodes are present in your study.}

\item{component_node}{Default 1. You can change it if more componentResults nodes are present in your study.}
}
\value{
list with
\itemize{
\item{"meta"}{General info like study ID or API version}
\item{"data"}{Tibble with all meta data, column \code{raw} stores the results for each entry in text format. This can be parsed for example with \code{jsonlite} or \code{jspsychread} packages}
}
}
\description{
Turns metadata and data into a single table
}
