% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{raw_json}
\alias{raw_json}
\title{Transforms raw data to JSON list}
\usage{
raw_json(d)
}
\arguments{
\item{d}{Tibble with data obtained with \code{process_results}}
}
\value{
Tibble with new list column \code{data_json} or a list with \code{data} tibble (see above)
}
\description{
Takes raw results stored in \code{data_raw} and adds new column \code{data_json} containing data in lists created by \code{jsonlite} package.
The package \code{jsonlite} must be installed for this function to work.
It is expected that the parameter \code{d} is the dataset obtained with \code{process_results},
but the dataset can be simpler, because only the column \code{data_raw} is used.
}
\details{
Function \code{process_results} returns a list (let's assume called \code{results}),
\code{raw_json} is intended to be used in tibble pipeline, i.e. on the \code{result$data} tibble.
To simplify the process, the function detects, whether it is provided
with a list containing a tibble called \code{data}. In that case,
it processes the \code{data} tibble, puts it back into the original list and returns it.
}
\examples{
\dontrun{
demo <- prepare_demo()
r <- process_results(demo$meta, demo$results)
# working with tibbles
rdata <- r$data |> raw_json()
# shortcut for the results list also works
r <- r |> raw_json()
clean_demo()
}
}
