% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_connection}
\alias{test_connection}
\title{Test JATOS connection}
\usage{
test_connection(jc)
}
\arguments{
\item{jc}{JATOS connection info, e.g. from \code{define_connection}}
}
\value{
HTTP response object returned by \code{httr2::req_perform()}
}
\description{
Test JATOS connection
}
\examples{
\dontrun{
cc <- define_connection(
  "https://www.myjatosinstance.org",
  "Bearer jap_xXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXxXx"
)
test_connection(cc)
}
}
